# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Project-Hub",
    "description": "API Specification for Project-Hub",
    "license": {
      "name": "MIT Licence",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "paths": {
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "description": "Login endpoint",
        "operationId": "loginUser",
        "requestBody": {
          "description": "new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token"
                  ],
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Wrong username or password"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "description": "Returns a single user based on the supplied id.\n",
        "operationId": "findUserById",
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Please log-in to access this API-function."
          },
          "404": {
            "description": "No user available, that matches the specified ID."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "description": "Updates a single user based on the supplied id.  \nThe supplied id has to be equal the id inside the JWT Token to access this request.  \n",
        "operationId": "updateUserById",
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "updated user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteUser"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "user successfully updated"
          },
          "400": {
            "description": "malformed request"
          },
          "401": {
            "description": "Unauthorized. Please log-in to access this API-function"
          },
          "403": {
            "description": "Forbidden. Not enough permissions!"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "description": "Deletes a single user based on the supplied id.  \nThe supplied id has to be equal the id inside the JWT Token to access this request.  \nThis call trigers a chain of events\n- delete the specefied user\n- remove the user from all projects he was part of\n- delete all projects where the user was the only member\n",
        "operationId": "deleteUser",
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "user successfully deleted"
          },
          "401": {
            "description": "Unauthorized. Please log-in to access this API-function"
          },
          "403": {
            "description": "Forbidden. Not enough permissions!"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/widgets/{id}": {
      "get": {
        "tags": [
          "widgets"
        ],
        "description": "Returns a widget based on the supplied ID.  \nThe interacting user can only use IDs of widgets that belong to projects he is part of.  \n",
        "operationId": "findWidgetByID",
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of widget to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "widget response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Please log-in to access this API-function"
          },
          "403": {
            "description": "Forbidden. Not enough permissions!"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "widgets"
        ],
        "description": "Edit a existing widget based on the supplied ID.  \nThe interacting user can only use IDs of widgets that belong to projects he is part of.  \nThis endpoint only executes if the interacting user has \"write\" permissions on the project this widget belongs to.  \n",
        "operationId": "updateWidgetById",
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of widget to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "updated widget",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWidget"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "widget successfully updated"
          },
          "400": {
            "description": "malformed request"
          },
          "401": {
            "description": "Unauthorized. Please log-in to access this API-function."
          },
          "403": {
            "description": "Forbidden. Not enough permissions!"
          },
          "404": {
            "description": "No Event available, that matches the specified ID."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "widgets"
        ],
        "description": "Delete a existing widget based on the supplied ID.  \nThe interacting user can only use IDs of widgets that belong to projects he is part of.  \nThis endpoint only executes if the interacting user has \"write\" permissions on the project this widget belongs to. \n",
        "operationId": "deleteWidget",
        "security": [
          {
            "user_auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of widget to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Widget successfully deleted."
          },
          "401": {
            "description": "Unauthorized. Please log-in to access this API-function"
          },
          "403": {
            "description": "Forbidden. Not enough permissions!"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "entryDate": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          }
        }
      },
      "WriteUser": {
        "allOf": [
          {
            "type": "object",
            "required": [
              "email",
              "username",
              "password"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          },
          {
            "$ref": "#/components/schemas/Profile"
          },
          {
            "type": "object",
            "properties": {
              "settings": {
                "type": "object"
              }
            }
          }
        ]
      },
...
      "User": {
        "allOf": [
          {
            "type": "object",
            "required": [
              "id",
              "username",
              "email"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          {
            "$ref": "#/components/schemas/Profile"
          },
          {
            "type": "object",
            "properties": {
              "settings": {
                "type": "object"
              }
            }
          }
        ]
      },
...
      "UpdateWidget": {
        "type": "object",
        "required": [
          "xPos",
          "yPos",
          "height",
          "width",
          "data"
        ],
        "properties": {
          "xPos": {
            "type": "number"
          },
          "yPos": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "width": {
            "type": "number"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "Widget": {
        "allOf": [
          {
            "type": "object",
            "required": [
              "id",
              "type"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            }
          },
          {
            "$ref": "#/components/schemas/UpdateWidget"
          }
        ]
      }
    },
    "securitySchemes": {
      "user_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}